Fix: NO_CAST.INTEGER_OVERFLOW
Problem:
The value of an arithmetic expression 'value << state->bits' is a subject to overflow because its operands are not cast to a larger data type before performing arithmetic
Solution: Cast value to unsigned long before the left shift to prevent signed integer overflow during shift evaluation.

+++ zlib/inflate.c
@@ -229,7 +229,7 @@ int ZEXPORT inflatePrime(z_streamp strm, int bits, int value)
     }
     if (bits > 16 || state->bits + bits > 32) return Z_STREAM_ERROR;
     value &= (1L << bits) - 1;
-    state->hold += value << state->bits;
+    state->hold += (unsigned long)value << state->bits;
     state->bits += bits;
     return Z_OK;
 }
